requirements.txt
scrapy
ipdb
service_identity
beautifulsoup
pytest
mock
sqlalchemy
alembic
mysql-python
pylint
pyyaml
gitpython
pygithub
